+++ src/transform.h
@@ -22,7 +22,7 @@
 @interface Transform: Node {
     double position[3], orientation[9];
     double translation[3], rotation[9];
-    float matrix[16];
+    double matrix[16];
 
     int transform;
 }
